Create Set_IC7300_date_time.py
Update to original code that set time. This version sets both date and time. The math to pick a time one minute in the future should even work if the date rolls over. There are three constants near the top of the code to set the baud rate and serial port for the ICOM IC-7300. The third constant chooses weather to set the clock to local time or UTC. There is a VERBOSE constant that can be set False to have no print out during the run of the code.